Backend: HKDF key derivation, sync ops, AES balance path, atomic rotation
- Derive confidential keys in one place (keys::derive_account_keys /
  derive_auditor_elgamal) with solana-zk-sdk's current HKDF-SHA512 scheme.
  The deprecated SHA3 derivation is opt-in via LEGACY_KDF=1 for devnet
  accounts provisioned before the migration; the crate-wide
  allow(deprecated) is gone.
- Every operation module is now a plain synchronous function on the
  blocking RPC client; handlers run them on tokio's blocking pool instead
  of spinning up a runtime per request. The unused TransferProgress /
  ProgressSink plumbing is removed.
- Withdraw and apply-pending read the available balance from the owner's
  AES-encrypted copy (what that field is for) instead of a 32-bit
  discrete-log search; view.rs propagates decryption failures instead of
  rendering a zero, with a hint pointing at LEGACY_KDF.
- Amounts are validated: non-finite, non-positive, sub-unit and oversized
  values are rejected with 400 before anything is built or paid for.
- Auditor-key rotation generates the key in memory, commits on-chain, and
  only then persists the keypair file, pointer and registry, so a failed
  RPC step leaves nothing on disk to wedge the next attempt.
- /transfer always requires the sender's owner token; the agent-proposal
  escape hatch and the origin_agent_proposal_id field are removed.
- /state reports the fee payer, the cluster derived from the RPC URL, and
  the confidential-transfer config read off the mint itself.
- Stale comments (SDK versions, "frontend-enforced" view permissions,
  pre-V1 multi-transaction flow) corrected; dead aliases and helpers removed.
- Unit tests for key derivation, keypair persistence, auth tokens/roles,
  amount parsing and cluster labelling.

+++ rust-service/src/auditor.rs
@@ -5,8 +5,8 @@
 //! Decryption fails safely (returns `Ok(None)`) if the ciphertext was
 //! encrypted to a different auditor generation's pubkey than the one whose
 //! secret is supplied here — the discrete log search simply finds no match.
-//! This is what gives CB-07's "wrong generation fails safely" requirement for
-//! free, with no special-case check needed.
+//! This is what makes "wrong generation fails safely" hold with no
+//! special-case check needed.
 
 use crate::types::*;
 use solana_zk_sdk::encryption::elgamal::{ElGamalCiphertext, ElGamalKeypair};
@@ -38,7 +38,7 @@ pub fn decrypt_auditor_amount(
     let lo_amount = lo_ct.decrypt_u32(auditor_elgamal.secret());
     let hi_amount = hi_ct.decrypt_u32(auditor_elgamal.secret());
     match (lo_amount, hi_amount) {
-        (Some(lo), Some(hi)) => Ok(Some((lo as u64) + ((hi as u64) << 16))),
+        (Some(lo), Some(hi)) => Ok(Some(lo + (hi << 16))),
         _ => Ok(None),
     }
 }
